compile -> implementation
compile is deprecated. This replaces all the usages of compile with
implementation.
This means that previously certain modules exposed their dependencies as
part of that module's API. This makes this no longer the case.
This also means that motif no longer exposes dagger as part of its API
and must be declared as a separate dependency.