Fix wrong method names generation
Context:
There is an edge case where motif would generate dependencies method names of inner classes with a "$" in the method name which could lead to compilation errors.

This simple one line fix will make sure that inner classes that have not been fully resolved by the compiler will be resolved and update their fullName property which is effecting the simpleName value used to generate the method names.

The logic in the java compiler to mutate the fullName property lives here:
https://github.com/google/error-prone-javac/blob/a53d069bbdb2c60232ed3811c19b65e41c3e60e0/src/jdk.compiler/share/classes/com/sun/tools/javac/code/Symtab.java#L618